Module ceres_liquidity_locker::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Structs

DefaultForAuthorityAccount

DefaultForFeesOptionOneAccount

DefaultForFeesOptionTwoAccount

DefaultForOptionTwoCeresAmount

DefaultForPalletStorageVersion

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Functions

__type_value_for_default_for_authority_account

__type_value_for_default_for_fees_option_one_account

__type_value_for_default_for_fees_option_two_account

__type_value_for_default_for_option_two_ceres_amount

__type_value_for_default_for_pallet_storage_version

Type Definitions

AccountIdOf

AssetIdOf

AuthorityAccount

Account which has permissions for changing CERES amount fee

FeesOptionOneAccount

Account for collecting fees from Option 1

FeesOptionTwoAccount

Account for collecting fees from Option 2

FeesOptionTwoCeresAmount

Amount of CERES for locker fees option two

LockerData

Contains data about lockups for each account

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PalletStorageVersion

Pallet storage version

Timestamp